Bytte extension från .mjs till .js för controller.mjs
Bröt även ut det mesta som har med frontend och webben att göra från controller.js till index.js.

Skapade separat fil för PageReader.

Allmän refaktorisering samt små buggfixar
